# Cache conditions when applying variables to config