--- a/internal/pkg/agent/transpiler/ast.go
+++ b/internal/pkg/agent/transpiler/ast.go
@@ -238,13 +238,14 @@ func (d *Dict) sort() {
 
 // Key represents a Key / value pair in the dictionary.
 type Key struct {
-	name  string
-	value Node
+	name      string
+	value     Node
+	condition *eql.Expression
 }
 
 // NewKey creates a new key with provided name node pair.
 func NewKey(name string, val Node) *Key {
-	return &Key{name, val}
+	return &Key{name: name, value: val}
 }
 
 func (k *Key) String() string {
@@ -334,11 +335,18 @@ func (k *Key) Apply(vars *Vars) (Node, error) {
 		case *BoolVal:
 			return k, nil
 		case *StrVal:
-			cond, err := eql.Eval(v.value, vars, true)
+			var err error
+			if k.condition == nil {
+				k.condition, err = eql.New(v.value)
+				if err != nil {
+					return nil, fmt.Errorf(`invalid condition "%s": %w`, v.value, err)
+				}
+			}
+			cond, err := k.condition.Eval(vars, true)
 			if err != nil {
 				return nil, fmt.Errorf(`condition "%s" evaluation failed: %w`, v.value, err)
 			}
-			return &Key{k.name, NewBoolVal(cond)}, nil
+			return &Key{name: k.name, value: NewBoolVal(cond)}, nil
 		}
 		return nil, fmt.Errorf("condition key's value must be a string; received %T", k.value)
 	}
@@ -349,7 +357,7 @@ func (k *Key) Apply(vars *Vars) (Node, error) {
 	if v == nil {
 		return nil, nil
 	}
-	return &Key{k.name, v}, nil
+	return &Key{name: k.name, value: v}, nil
 }
 
 // Processors returns any attached processors, because of variable substitution.