--- a/internal/pkg/agent/transpiler/ast_test.go
+++ b/internal/pkg/agent/transpiler/ast_test.go
@@ -11,6 +11,8 @@ import (
 
 	"github.com/elastic/elastic-agent-libs/mapstr"
 
+	"github.com/elastic/elastic-agent/internal/pkg/eql"
+
 	"github.com/google/go-cmp/cmp"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -1106,6 +1108,29 @@ func TestLookup(t *testing.T) {
 			}
 		})
 	}
+}
+
+func TestCondition(t *testing.T) {
+	vars := mustMakeVars(map[string]interface{}{
+		"other": map[string]interface{}{
+			"data": "info",
+		}})
+
+	input := NewKey("condition", NewStrVal("${other.data} == 'info'"))
+	expected := NewKey("condition", NewBoolVal(true))
+
+	// the condition string hasn't been parsed yet
+	assert.Nil(t, input.condition)
+
+	output, err := input.Apply(vars)
+	require.NoError(t, err)
+	assert.Equal(t, expected, output)
+
+	// check if the condition was parsed and cached
+	assert.NotNil(t, input.condition)
+	condition, err := eql.New(input.value.Value().(string))
+	require.NoError(t, err)
+	assert.Equal(t, condition, input.condition)
 }
 
 func mustMakeVars(mapping map[string]interface{}) *Vars {